change type of images number

+++ src/image/oci/storage/image_store/image_store.c
@@ -2836,7 +2836,7 @@ out:
     return ret;
 }
 
-int image_store_get_images_number()
+size_t image_store_get_images_number()
 {
     size_t number = 0;
